backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "boto3>=1.42.14",
    "celery>=5.6.0",
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-aws>=1.1.0",
    "langchain-openai>=1.1.5",
    "langgraph>=1.0.5",
    "neo4j>=6.0.3",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
